Moved AbstractOverridableDTO to commons from commons-mongo
