+++ commons-core/src/main/java/com/fincity/saas/commons/core/document/Action.java
@@ -1,6 +1,6 @@
 package com.fincity.saas.commons.core.document;
 
-import com.fincity.saas.commons.mongo.model.AbstractOverridableDTO;
+import com.fincity.saas.commons.model.dto.AbstractOverridableDTO;
 import com.fincity.saas.commons.util.CloneUtil;
 import com.fincity.saas.commons.util.DifferenceApplicator;
 import com.fincity.saas.commons.util.DifferenceExtractor;
