+++ commons-core/src/main/java/com/fincity/saas/commons/core/document/EventAction.java
@@ -1,7 +1,7 @@
 package com.fincity.saas.commons.core.document;
 
 import com.fincity.saas.commons.core.model.EventActionTask;
-import com.fincity.saas.commons.mongo.model.AbstractOverridableDTO;
+import com.fincity.saas.commons.model.dto.AbstractOverridableDTO;
 import com.fincity.saas.commons.util.CloneUtil;
 import com.fincity.saas.commons.util.DifferenceApplicator;
 import com.fincity.saas.commons.util.DifferenceExtractor;
