+++ commons-core/src/main/java/com/fincity/saas/commons/core/document/Storage.java
@@ -3,7 +3,7 @@ package com.fincity.saas.commons.core.document;
 import com.fincity.nocode.reactor.util.FlatMapUtil;
 import com.fincity.saas.commons.core.enums.StorageTriggerType;
 import com.fincity.saas.commons.core.model.StorageRelation;
-import com.fincity.saas.commons.mongo.model.AbstractOverridableDTO;
+import com.fincity.saas.commons.model.dto.AbstractOverridableDTO;
 import com.fincity.saas.commons.util.CloneUtil;
 import com.fincity.saas.commons.util.CommonsUtil;
 import com.fincity.saas.commons.util.DifferenceApplicator;
