+++ commons-core/src/main/java/com/fincity/saas/commons/core/document/Workflow.java
@@ -2,7 +2,7 @@ package com.fincity.saas.commons.core.document;
 
 import com.fincity.saas.commons.core.model.TriggerWorkflowStep;
 import com.fincity.saas.commons.core.model.WorkflowStep;
-import com.fincity.saas.commons.mongo.model.AbstractOverridableDTO;
+import com.fincity.saas.commons.model.dto.AbstractOverridableDTO;
 import com.fincity.saas.commons.util.CloneUtil;
 import java.io.Serial;
 import java.util.Map;
