+++ commons-mongo/src/main/java/com/fincity/saas/commons/mongo/document/AbstractFiller.java
@@ -3,7 +3,7 @@ package com.fincity.saas.commons.mongo.document;
 import java.util.Map;
 
 import com.fincity.nocode.reactor.util.FlatMapUtil;
-import com.fincity.saas.commons.mongo.model.AbstractOverridableDTO;
+import com.fincity.saas.commons.model.dto.AbstractOverridableDTO;
 import com.fincity.saas.commons.util.CloneUtil;
 import com.fincity.saas.commons.util.DifferenceApplicator;
 import com.fincity.saas.commons.util.DifferenceExtractor;
