+++ commons-mongo/src/main/java/com/fincity/saas/commons/mongo/document/AbstractFunction.java
@@ -2,7 +2,7 @@ package com.fincity.saas.commons.mongo.document;
 
 import java.util.Map;
 
-import com.fincity.saas.commons.mongo.model.AbstractOverridableDTO;
+import com.fincity.saas.commons.model.dto.AbstractOverridableDTO;
 import com.fincity.saas.commons.util.CloneUtil;
 import com.fincity.saas.commons.util.DifferenceApplicator;
 import com.fincity.saas.commons.util.DifferenceExtractor;
