+++ message/src/main/java/com/fincity/saas/message/oserver/core/document/Connection.java
@@ -1,6 +1,6 @@
 package com.fincity.saas.message.oserver.core.document;
 
-import com.fincity.saas.commons.mongo.model.AbstractOverridableDTO;
+import com.fincity.saas.commons.model.dto.AbstractOverridableDTO;
 import com.fincity.saas.commons.util.CloneUtil;
 import com.fincity.saas.message.oserver.core.enums.ConnectionSubType;
 import com.fincity.saas.message.oserver.core.enums.ConnectionType;
