+++ notification/pom.xml
@@ -40,7 +40,6 @@
         <package.i.nocode-flatmap-util-version>1.3.0</package.i.nocode-flatmap-util-version>
         <package.i.nocode-kirun-version>3.7.1</package.i.nocode-kirun-version>
         <package.i.saas-commons-jooq-version>2.0.0</package.i.saas-commons-jooq-version>
-        <package.i.saas-commons-mongo-version>2.0.0</package.i.saas-commons-mongo-version>
         <package.i.saas-commons-mq-version>2.0.0</package.i.saas-commons-mq-version>
         <package.i.saas-commons-security-version>2.0.0</package.i.saas-commons-security-version>
         <package.i.saas-commons-version>2.0.0</package.i.saas-commons-version>
@@ -90,11 +89,6 @@
         </dependency>
         <dependency>
             <groupId>com.fincity.saas</groupId>
-            <artifactId>commons-mongo</artifactId>
-            <version>${package.i.saas-commons-mongo-version}</version>
-        </dependency>
-        <dependency>
-            <groupId>com.fincity.saas</groupId>
             <artifactId>commons-mq</artifactId>
             <version>${package.i.saas-commons-mq-version}</version>
         </dependency>
@@ -102,6 +96,10 @@
             <groupId>com.fincity.saas</groupId>
             <artifactId>commons-security</artifactId>
             <version>${package.i.saas-commons-security-version}</version>
+        </dependency>
+        <dependency>
+            <groupId>com.mysql</groupId>
+            <artifactId>mysql-connector-j</artifactId>
         </dependency>
         <dependency>
             <groupId>com.playtika.reactivefeign</groupId>
@@ -201,10 +199,6 @@
             <artifactId>spring-cloud-starter-netflix-eureka-client</artifactId>
         </dependency>
         <dependency>
-            <groupId>com.mysql</groupId>
-            <artifactId>mysql-connector-j</artifactId>
-        </dependency>
-        <dependency>
             <groupId>jakarta.websocket</groupId>
             <artifactId>jakarta.websocket-api</artifactId>
             <version>${package.o.jakarta.websocket-api.version}</version>
@@ -289,41 +283,29 @@
                                     <inputSchema>${database.schema.name}</inputSchema>
                                     <forcedTypes>
                                         <forcedType>
-                                            <userType>com.fincity.saas.notification.model.request.NotificationChannel
-                                            </userType>
-                                            <converter>
-                                                com.fincity.saas.commons.jooq.convertor.jooq.converters.JSONtoClassConverter
-                                            </converter>
+                                            <userType>com.fincity.saas.notification.model.request.NotificationChannel</userType>
+                                            <converter>com.fincity.saas.commons.jooq.convertor.jooq.converters.JSONtoClassConverter</converter>
                                             <includeTypes>JSON</includeTypes>
                                             <genericConverter>true</genericConverter>
                                             <includeExpression>.*NOTIFICATION_CHANNEL</includeExpression>
                                         </forcedType>
                                         <forcedType>
-                                            <userType>com.fincity.saas.notification.model.message.channel.InAppMessage
-                                            </userType>
-                                            <converter>
-                                                com.fincity.saas.commons.jooq.convertor.jooq.converters.JSONtoClassConverter
-                                            </converter>
+                                            <userType>com.fincity.saas.notification.model.message.channel.InAppMessage</userType>
+                                            <converter>com.fincity.saas.commons.jooq.convertor.jooq.converters.JSONtoClassConverter</converter>
                                             <includeTypes>JSON</includeTypes>
                                             <genericConverter>true</genericConverter>
                                             <includeExpression>.*IN_APP_MESSAGE</includeExpression>
                                         </forcedType>
                                         <forcedType>
-                                            <userType>
-                                                com.fincity.saas.notification.model.response.NotificationErrorInfo
-                                            </userType>
-                                            <converter>
-                                                com.fincity.saas.commons.jooq.convertor.jooq.converters.JSONtoClassConverter
-                                            </converter>
+                                            <userType>com.fincity.saas.notification.model.response.NotificationErrorInfo</userType>
+                                            <converter>com.fincity.saas.commons.jooq.convertor.jooq.converters.JSONtoClassConverter</converter>
                                             <includeTypes>JSON</includeTypes>
                                             <genericConverter>true</genericConverter>
                                             <includeExpression>.*ERROR_INFO</includeExpression>
                                         </forcedType>
                                         <forcedType>
                                             <userType>java.util.Map</userType>
-                                            <converter>
-                                                com.fincity.saas.commons.jooq.convertor.jooq.converters.JSONtoClassConverter
-                                            </converter>
+                                            <converter>com.fincity.saas.commons.jooq.convertor.jooq.converters.JSONtoClassConverter</converter>
                                             <includeTypes>JSON</includeTypes>
                                             <genericConverter>true</genericConverter>
                                         </forcedType>
@@ -340,8 +322,7 @@
                                             <includeTypes>ENUM</includeTypes>
                                         </forcedType>
                                         <forcedType>
-                                            <userType>com.fincity.saas.notification.enums.NotificationDeliveryStatus
-                                            </userType>
+                                            <userType>com.fincity.saas.notification.enums.NotificationDeliveryStatus</userType>
                                             <enumConverter>true</enumConverter>
                                             <includeExpression>.*\.NOTIFICATION_DELIVERY_STATUS</includeExpression>
                                             <includeTypes>ENUM</includeTypes>
