+++ notification/src/main/java/com/fincity/saas/notification/oserver/core/document/Notification.java
@@ -1,6 +1,6 @@
 package com.fincity.saas.notification.oserver.core.document;
 
-import com.fincity.saas.commons.mongo.model.AbstractOverridableDTO;
+import com.fincity.saas.commons.model.dto.AbstractOverridableDTO;
 import com.fincity.saas.commons.util.CloneUtil;
 import com.fincity.saas.notification.enums.channel.NotificationChannelType;
 import com.fincity.saas.notification.model.template.NotificationTemplate;
