+++ notification/src/main/java/com/fincity/saas/notification/oserver/core/service/AbstractCoreService.java
@@ -1,7 +1,7 @@
 package com.fincity.saas.notification.oserver.core.service;
 
 import com.fincity.nocode.reactor.util.FlatMapUtil;
-import com.fincity.saas.commons.mongo.model.AbstractOverridableDTO;
+import com.fincity.saas.commons.model.dto.AbstractOverridableDTO;
 import com.fincity.saas.commons.security.feign.IFeignSecurityService;
 import com.fincity.saas.commons.service.CacheService;
 import com.fincity.saas.notification.feign.IFeignCoreService;
