+++ ui/src/main/java/com/fincity/saas/ui/document/Personalization.java
@@ -5,7 +5,7 @@ import java.util.Map;
 import org.springframework.data.mongodb.core.index.CompoundIndex;
 import org.springframework.data.mongodb.core.mapping.Document;
 
-import com.fincity.saas.commons.mongo.model.AbstractOverridableDTO;
+import com.fincity.saas.commons.model.dto.AbstractOverridableDTO;
 
 import lombok.Data;
 import lombok.EqualsAndHashCode;
